requirements.txt
torch>=2.0.0
transformers>=4.35.0
sae-lens>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.7.0
datasets>=2.14.0
jaxtyping>=0.2.28
jupyter>=1.0.0
ipykernel>=6.0.0
